% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pub_options}
\alias{pub_options}
\title{Set or get options for my package}
\usage{
pub_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
the default and modified options.
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
 host: host of tiddlywiki
 output: output for intermediate files
 author_max: Maximum number of authors/colleagues to download
 file_expired: days to intermediate files expired
 file_remove_max: maximum number of intermediate file to remove
 latest_literature: tiddler name for latest literature
}

\examples{
pub_options(host = "http://127.0.0.1:8080/")
}
